[df] Move GetBranchNames logic to RTTreeDS
