+++ tree/dataframe/src/RLoopManager.cxx
@@ -33,7 +33,6 @@
 #include "TROOT.h" // IsImplicitMTEnabled, gCoreMutex, R__*_LOCKGUARD
 #include "TTreeReader.h"
 #include "TTree.h" // For MaxTreeSizeRAII. Revert when #6640 will be solved.
-#include "ROOT/InternalTreeUtils.hxx" // GetTopLevelBranchNames
 
 #include "ROOT/RTTreeDS.hxx"
 
@@ -89,174 +88,6 @@ std::string &GetCodeToJit()
 {
    static std::string code;
    return code;
-}
-
-bool ContainsLeaf(const std::set<TLeaf *> &leaves, TLeaf *leaf)
-{
-   return (leaves.find(leaf) != leaves.end());
-}
-
-///////////////////////////////////////////////////////////////////////////////
-/// This overload does not check whether the leaf/branch is already in bNamesReg. In case this is a friend leaf/branch,
-/// `allowDuplicates` controls whether we add both `friendname.bname` and `bname` or just the shorter version.
-void InsertBranchName(std::set<std::string> &bNamesReg, ColumnNames_t &bNames, const std::string &branchName,
-                             const std::string &friendName, bool allowDuplicates)
-{
-   if (!friendName.empty()) {
-      // In case of a friend tree, users might prepend its name/alias to the branch names
-      const auto friendBName = friendName + "." + branchName;
-      if (bNamesReg.insert(friendBName).second)
-         bNames.push_back(friendBName);
-   }
-
-   if (allowDuplicates || friendName.empty()) {
-      if (bNamesReg.insert(branchName).second)
-         bNames.push_back(branchName);
-   }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-/// This overload makes sure that the TLeaf has not been already inserted.
-void InsertBranchName(std::set<std::string> &bNamesReg, ColumnNames_t &bNames, const std::string &branchName,
-                             const std::string &friendName, std::set<TLeaf *> &foundLeaves, TLeaf *leaf,
-                             bool allowDuplicates)
-{
-   const bool canAdd = allowDuplicates ? true : !ContainsLeaf(foundLeaves, leaf);
-   if (!canAdd) {
-      return;
-   }
-
-   InsertBranchName(bNamesReg, bNames, branchName, friendName, allowDuplicates);
-
-   foundLeaves.insert(leaf);
-}
-
-void ExploreBranch(TTree &t, std::set<std::string> &bNamesReg, ColumnNames_t &bNames, TBranch *b,
-                          std::string prefix, std::string &friendName, bool allowDuplicates)
-{
-   // We want to avoid situations of overlap between the prefix and the
-   // sub-branch name that might happen when the branch is composite, e.g.
-   // prefix=reco_Wdecay2_from_tbar_4vect_NOSYS.fCoordinates.
-   // subBranchName=fCoordinates.fPt
-   // which would lead to a repetition of fCoordinates in the output branch name
-   // Boundary to search for the token before the last dot
-   auto prefixEndingDot = std::string::npos;
-   if (!prefix.empty() && prefix.back() == '.')
-      prefixEndingDot = prefix.size() - 2;
-   std::string lastPrefixToken{};
-   if (auto prefixLastRealDot = prefix.find_last_of('.', prefixEndingDot); prefixLastRealDot != std::string::npos)
-      lastPrefixToken = prefix.substr(prefixLastRealDot + 1, prefixEndingDot - prefixLastRealDot);
-
-   for (auto sb : *b->GetListOfBranches()) {
-      TBranch *subBranch = static_cast<TBranch *>(sb);
-      auto subBranchName = std::string(subBranch->GetName());
-      auto fullName = prefix + subBranchName;
-
-      if (auto subNameFirstDot = subBranchName.find_first_of('.'); subNameFirstDot != std::string::npos) {
-         // Concatenate the prefix to the sub-branch name without overlaps
-         if (!lastPrefixToken.empty() && lastPrefixToken == subBranchName.substr(0, subNameFirstDot))
-            fullName = prefix + subBranchName.substr(subNameFirstDot + 1);
-      }
-
-      std::string newPrefix;
-      if (!prefix.empty())
-         newPrefix = fullName + ".";
-
-      ExploreBranch(t, bNamesReg, bNames, subBranch, newPrefix, friendName, allowDuplicates);
-
-      auto branchDirectlyFromTree = t.GetBranch(fullName.c_str());
-      if (!branchDirectlyFromTree)
-         branchDirectlyFromTree = t.FindBranch(fullName.c_str()); // try harder
-      if (branchDirectlyFromTree)
-         InsertBranchName(bNamesReg, bNames, std::string(branchDirectlyFromTree->GetFullName()), friendName,
-                          allowDuplicates);
-
-      if (bNamesReg.find(subBranchName) == bNamesReg.end() && t.GetBranch(subBranchName.c_str()))
-         InsertBranchName(bNamesReg, bNames, subBranchName, friendName, allowDuplicates);
-   }
-}
-
-void GetBranchNamesImpl(TTree &t, std::set<std::string> &bNamesReg, ColumnNames_t &bNames,
-                               std::set<TTree *> &analysedTrees, std::string &friendName, bool allowDuplicates)
-{
-   std::set<TLeaf *> foundLeaves;
-   if (!analysedTrees.insert(&t).second) {
-      return;
-   }
-
-   const auto branches = t.GetListOfBranches();
-   // Getting the branches here triggered the read of the first file of the chain if t is a chain.
-   // We check if a tree has been successfully read, otherwise we throw (see ROOT-9984) to avoid further
-   // operations
-   if (!t.GetTree()) {
-      std::string err("GetBranchNames: error in opening the tree ");
-      err += t.GetName();
-      throw std::runtime_error(err);
-   }
-   if (branches) {
-      for (auto b : *branches) {
-         TBranch *branch = static_cast<TBranch *>(b);
-         const auto branchName = std::string(branch->GetName());
-         if (branch->IsA() == TBranch::Class()) {
-            // Leaf list
-            auto listOfLeaves = branch->GetListOfLeaves();
-            if (listOfLeaves->GetEntriesUnsafe() == 1) {
-               auto leaf = static_cast<TLeaf *>(listOfLeaves->UncheckedAt(0));
-               InsertBranchName(bNamesReg, bNames, branchName, friendName, foundLeaves, leaf, allowDuplicates);
-            }
-
-            for (auto leaf : *listOfLeaves) {
-               auto castLeaf = static_cast<TLeaf *>(leaf);
-               const auto leafName = std::string(leaf->GetName());
-               const auto fullName = branchName + "." + leafName;
-               InsertBranchName(bNamesReg, bNames, fullName, friendName, foundLeaves, castLeaf, allowDuplicates);
-            }
-         } else if (branch->IsA() == TBranchObject::Class()) {
-            // TBranchObject
-            ExploreBranch(t, bNamesReg, bNames, branch, branchName + ".", friendName, allowDuplicates);
-            InsertBranchName(bNamesReg, bNames, branchName, friendName, allowDuplicates);
-         } else {
-            // TBranchElement
-            // Check if there is explicit or implicit dot in the name
-
-            bool dotIsImplied = false;
-            auto be = dynamic_cast<TBranchElement *>(b);
-            if (!be)
-               throw std::runtime_error("GetBranchNames: unsupported branch type");
-            // TClonesArray (3) and STL collection (4)
-            if (be->GetType() == 3 || be->GetType() == 4)
-               dotIsImplied = true;
-
-            if (dotIsImplied || branchName.back() == '.')
-               ExploreBranch(t, bNamesReg, bNames, branch, "", friendName, allowDuplicates);
-            else
-               ExploreBranch(t, bNamesReg, bNames, branch, branchName + ".", friendName, allowDuplicates);
-
-            InsertBranchName(bNamesReg, bNames, branchName, friendName, allowDuplicates);
-         }
-      }
-   }
-
-   // The list of friends needs to be accessed via GetTree()->GetListOfFriends()
-   // (and not via GetListOfFriends() directly), otherwise when `t` is a TChain we
-   // might not recover the list correctly (https://github.com/root-project/root/issues/6741).
-   auto friendTrees = t.GetTree()->GetListOfFriends();
-
-   if (!friendTrees)
-      return;
-
-   for (auto friendTreeObj : *friendTrees) {
-      auto friendTree = ((TFriendElement *)friendTreeObj)->GetTree();
-
-      std::string frName;
-      auto alias = t.GetFriendAlias(friendTree);
-      if (alias != nullptr)
-         frName = std::string(alias);
-      else
-         frName = std::string(friendTree->GetName());
-
-      GetBranchNamesImpl(*friendTree, bNamesReg, bNames, analysedTrees, frName, allowDuplicates);
-   }
 }
 
 void ThrowIfNSlotsChanged(unsigned int nSlots)
@@ -382,18 +213,6 @@ struct RCallCleanUpTask {
 } // namespace RDF
 } // namespace Detail
 } // namespace ROOT
-
-///////////////////////////////////////////////////////////////////////////////
-/// Get all the branches names, including the ones of the friend trees
-ColumnNames_t ROOT::Internal::RDF::GetBranchNames(TTree &t, bool allowDuplicates)
-{
-   std::set<std::string> bNamesSet;
-   ColumnNames_t bNames;
-   std::set<TTree *> analysedTrees;
-   std::string emptyFrName = "";
-   GetBranchNamesImpl(t, bNamesSet, bNames, analysedTrees, emptyFrName, allowDuplicates);
-   return bNames;
-}
 
 ROOT::Detail::RDF::RLoopManager::RLoopManager(const ROOT::Detail::RDF::ColumnNames_t &defaultColumns)
    : fDefaultColumns(defaultColumns),
